Rework Partition memory management.
Copy construction of Partition objects is now delayed and delegated to
the job itself. As a result, the "parent" Partition, which is used as an
argument to the Partition copy constructor, must continue to exist until
all jobs that require it are completed. Object ownership is maintained
using shared pointers. To enforce the use of shared pointers,
Partition's constructors are now private, and factory functions which
return shared pointers are used instead.