Add an extra chance at optimization.
Previously, if the Newton-Raphson step took the next branch length
negative, iteration would halt. This change allows it an extra chance in
the hopes that evaluating the derivatives at a small positive branch
length will result in a positive step toward an optimum. Iteration will
halt if another negative branch length is encountered.