Add some unit tests.
The wanderer vs. other methods section currently fails because the
tree tables are static member variables of the Wanderer class, and
thus any wanderer created after the first thinks the initial tree has
already been evaluated and throws an exception. Making the tables
non-static members of the Authority class, which is on the to-do list,
will solve this problem.