Rework guru multiple starting tree constructor.
Responsibility for synchronizing tip node/partition data indices is
now on the caller, which libptpll's ParseMultiNewick() will do for us.
The starting trees are now also sorted by descending log-likelihood
before being added to the queue so that the search proceeds
predictably no matter the order of the starting trees.